refactor constants.go to config
