+++ internal/docker/docker.go
@@ -17,7 +17,7 @@ package docker
 import (
 	"context"
 	"fmt"
-	"github.com/googleapis/librarian/internal/constants"
+	"github.com/googleapis/librarian/internal/config"
 	"log/slog"
 	"os"
 	"os/exec"
@@ -109,13 +109,13 @@ func (c *Docker) GenerateLibrary(apiRoot, output, generatorInput, libraryID stri
 	commandArgs := []string{
 		"--api-root=/apis",
 		"--output=/output",
-		fmt.Sprintf("--%s=/%s", constants.GeneratorInputDir, constants.GeneratorInputDir),
+		fmt.Sprintf("--%s=/%s", config.GeneratorInputDir, config.GeneratorInputDir),
 		fmt.Sprintf("--library-id=%s", libraryID),
 	}
 	mounts := []string{
 		fmt.Sprintf("%s:/apis", apiRoot),
 		fmt.Sprintf("%s:/output", output),
-		fmt.Sprintf("%s:/%s", generatorInput, constants.GeneratorInputDir),
+		fmt.Sprintf("%s:/%s", generatorInput, config.GeneratorInputDir),
 	}
 	return c.runDocker(CommandGenerateLibrary, mounts, commandArgs)
 }
@@ -182,12 +182,12 @@ func (c *Docker) Configure(apiRoot, apiPath, generatorInput string) error {
 	}
 	commandArgs := []string{
 		"--api-root=/apis",
-		fmt.Sprintf("--%s=/%s", constants.GeneratorInputDir, constants.GeneratorInputDir),
+		fmt.Sprintf("--%s=/%s", config.GeneratorInputDir, config.GeneratorInputDir),
 		fmt.Sprintf("--api-path=%s", apiPath),
 	}
 	mounts := []string{
 		fmt.Sprintf("%s:/apis", apiRoot),
-		fmt.Sprintf("%s:/%s", generatorInput, constants.GeneratorInputDir),
+		fmt.Sprintf("%s:/%s", generatorInput, config.GeneratorInputDir),
 	}
 	return c.runDocker(CommandConfigure, mounts, commandArgs)
 }
